RDS wrapup
- Store `alarm_endpoints` on each `AppRegion` object, so that triggers
  outside of the `AlarmFactory` can reference them too
- Refactor `TriggerFactory` to a superclass, provide an alternative
  implementation for RDS triggers.
- Adding a `multi_region` test suite for verifying the RDS `global`
  flag: looks good (after a few fixes).
